Make the access_key_id and access_key_data optional
Summary:
Changed the access_key_id and access_key_data in `AWSContainerService` into optional parameters. With this change,
1. It won't affect existing functionalities. As you can tell, to make them optional, I  switched the order of the parameters. I also switched where this class was used, including the config.yml file. So whatever worked before should still work.
2. One additional benefit is I can get rid of the lines with `access_key_id` and `access_key_data` inside of the config.xml, as long as there are environment variables with these two credentials being set. According to this doc https://boto3.amazonaws.com/v1/documentation/api/1.10.44/guide/configuration.html#configuring-credentials, the boto client will look for the credentials in several places, and one of them is the environment variables.

The CloudBridge stuff I'm working on need to make sure that there are no manual steps on the advertiser side. The way it works is that we populate the config.yml file with all the configuration, and upload the config.yml to s3. But since there're the AWS credentials, we don't want to upload a file with credentials inside to S3. The advertiser's environment will have the AWS credentials as environment variable, so with the #2 benefit mentioned above, I could get rid of the lines with the credential and upload this file to S3. Then the advertiser side will read the AWS credentials from environment variable instead of from the config.yml file.

Reviewed By: peking2

Differential Revision: D28022084

fbshipit-source-id: 12e5aaf376f562c26d2edcb8737ac171c823abb3