Refactor stop_task in ECSGateway
Summary:
This diff is to redefine the behavior of ECSGateway. Previously stop_task returns Dict[stry, Any] which was not appropriate as the main responsibility of a gateway is to transform data from external format to an entity. As stop_task is an aysnc operation, we can assume the task will be killed eventually if no exceptions thrown.

Next: need to redefine the behavior of AWSContainerService.

Differential Revision: D28920817

fbshipit-source-id: d9f71e229afa7e240101d9962fb56875c37713bf