Refactor stop_task in ECSGateway
Summary:
This diff is to redefine the behavior of ECSGateway. Previously stop_task returns Dict[stry, Any] which was not appropriate as the main responsibility of a gateway is to transform data from external format to an entity. Now we return ContainerInstance instead.

Next: need to redefine the behavior of AWSContainerService.

Differential Revision: D28920817

fbshipit-source-id: dac9de8001fd4fc1bcc294462e1dfec9ed300b74