Refactor stop_task in ECSGateway (#25)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/25

This diff is to redefine the behavior of ECSGateway. Previously stop_task returns Dict[stry, Any] which was not appropriate as the main responsibility of a gateway is to transform data from external format to an entity. As stop_task is an aysnc operation, we can assume the task will be killed eventually if no exceptions thrown.

Next: need to redefine the behavior of AWSContainerService.

Reviewed By: wenhaizhu

Differential Revision: D28920817

fbshipit-source-id: 6b516093a73fb8243339c2e21fffa1026baabd98