Use a list of subnets instead of one subnet when run tasks (#42)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/42

This diff changed the code on the ContainerService level. It will accept subnet ids as a string (e.g. "subnet_id0, subnetid1") and pass in a "List" of subnet ids into the ecs gateway.

With this change, we don't need to change any config.yml file, and if we want to use two subnets, just put the subnet ids into the "subnet" in config.yml.

Reviewed By: peking2

Differential Revision: D29105794

fbshipit-source-id: 4b65af4207273c253f40cf03e27b2dd56daab659