Added Get Status to OWDL Driver (#65)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/65

Added get_status to OWDLDriver, refactored to return OWDLWorkflowInstance

Status updates are based on the first container with a non-COMPLETED status. On a STARTED, UNKNOWN, or FAILED container, the status of both the state and the workflow will be marked the same as the container. If all of the containers are COMPLETED, then the state is marked as COMPLETED and the workflow status remains unchanged.

If there are containers of multiple statuses, the priority is FAILED > STARTED > UNKNOWN > COMPLETED

Reviewed By: peking2

Differential Revision: D29468685

fbshipit-source-id: 1c776b84c38a67c5fe2fa68247771d4133aaa6e5