Add env override in ecs gateway (#87)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/87

## Context
Pid is migrating to OneDocker and it relies on dynamic env overrides. This diff will add env overrides to ecs gateway so upstream users can pass any env they want to override in the container.

## Summary
Add envs to run_task in ecs_gateway

The format of env should be: {"name": "", "value": ""} and we pass a list of envs to run_task

## Next
1. Add env override support in container service
2. Add env override support in onedocker service.

Differential Revision: D29889149

fbshipit-source-id: 39b5ddbfb4ca4f094e8410535063677557a77d6f