Add subnet entity in fbpcs (#91)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/91

## Context
We plan to fetch all possible subnets in one specific vpc in our stress test script. So we need this subnet entity to store the subnet information.

## Summary
The subnet entity will container id, az and tags.
Please check the response from ec2 client to see if the fields are reasonable or if we need other variables

https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ec2.html#EC2.Client.describe_subnets

## Next
1. Implement subnet mapper
2. Implement describe subnets in ec2 gateway

Reviewed By: peking2

Differential Revision: D29952426

fbshipit-source-id: 64fa1bc166699ca62db36a6b0189cbe8e3b42444