Define MetricsGetter interface (#111)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/111

This diff is to define MetricsGetter and metrics decorators will depend on it. In order to leverage metrics decorators, a class has to implement MetricsGetter. It's more Java like and strongly typed in order to address previous comment that it's too implicit that the class has to have  metrics field.

Reviewed By: ajaybhargavb

Differential Revision: D30240828

fbshipit-source-id: 794252b28409b35a4053a9034593bd8c844184d1