Rename FBPCS to FBPCP in README
Summary:
We decided to rename some projects. Basically
FBPCS -> FBPCP
FBPMP -> FBPCS

We originally picked up fbpcp, but we had to change it to fbpcs due to some concerns. However, it doesn't make sense that a platform layer is built on top of a service layer, so we want to fix them altogether.

zhuang-93 will work on following items
- Release FBPCP and deprecate FBPCS
- Move to fbpcp repo

Reviewed By: gorel

Differential Revision: D30259784

fbshipit-source-id: d74e94114a6036fb2141f4bb9cf86197ad9e4a11