Rename FBPCS to FBPCP in README (#116)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FBPCS/pull/116

We decided to rename some projects. Basically
FBPCS -> FBPCP
FBPMP -> FBPCS

We originally picked up fbpcp, but we had to change it to fbpcs due to some concerns. However, it doesn't make sense that a platform layer is built on top of a service layer, so we want to fix them altogether.

zhuang-93 will work on following items
- Release FBPCP and deprecate FBPCS
- Move to fbpcp repo

Reviewed By: gorel

Differential Revision: D30259784

fbshipit-source-id: 5e1efd4acad40408f969cd50f2546131524feda8