Add function to get the container definition (#148)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/148

**What:**
2. Add gateway to describe the task definition from aws.
3. Add mapper function to map the aws response to pce entity.
4. Add integration test and unit test.

**RFC:**
We have a circular dependency issue if the entity don't live in the same package with the aws gateway and mapper classes. I want to get some advise here. Which is better? Move the `container_definition.py` to oss or we move all gateway and mapper related to the `container_definition` out of oss?

Differential Revision: D30812929

fbshipit-source-id: 61fa2b085cc039db358fc2b1dc034877adb034b6