Deprecate list_tasks
Summary:
list_tasks is not defined in ContainerService. We shouldn't add custom methods. It will ruin the defined interface. Looks like it's introduced in D25739819.

As everyone has onboarded to OneDocker and we provide timeout from OneDocker, we don't have to manually kill the containers.

Also given the conversation we had recently, this aws-cli tool is rarely used now.

Differential Revision: D31042674

fbshipit-source-id: fc2c733366e7f0cdcf62894eda4782a6ea8416ca