Remove region from BillingService and CostExplorerGateway
Summary:
**What**
- remove region from the input variable from BillingService and CostExplorerGateway
- make region optional in AWSGateway

**Why**
For billing service and CostExplorerGateway, region is not used. Boto client for costExplorer doesn't need region as input. It will always return all regions unless desire regions are listed in filter. So having this region in the constructor might confuse user. Also make this region optional in AWSGateway because there are some other boto clients doesn't need this region (e.g. S3) and we will want to clean it up as well.

Differential Revision: D31059590

fbshipit-source-id: 5647174c5ff3badbdcfdff05a7e32a57be77afd6