Deprecate list_tasks (#152)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/152

list_tasks is not defined in ContainerService. We shouldn't add custom methods. It will ruin the defined interface. Looks like it's introduced in D25739819.

As everyone has onboarded to OneDocker and we provide timeout from OneDocker, we don't have to manually kill the containers.

Also given the conversation we had recently, this aws-cli tool is rarely used now.

Reviewed By: jliu0501

Differential Revision: D31042674

fbshipit-source-id: 879cb6c8054e133e4ff840bb1903e884e5a2a9b1