Remove storage_svc from MPCService's constructor (#160)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/160

I just noticed that MPCService requires storage service. It doesn't seem right. I think it's introduced because we needed it for ip exchange. This feature has been deprecated long time ago. Unfortunately, PL depends on it in a wrong way. PL should depend on MPCService, shouldn't depend on a dependency of MPCService, so we need to fix them.

Reviewed By: jrodal98

Differential Revision: D31226903

fbshipit-source-id: 8349a481c63a4473e2608bdb7548827f2418c7e2