Add AWS services to ServiceName (#158)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/158

After offline discussion about the design, we decide that these two enum class ServiceName and RegionName is not a good way to keep up with the potential name change. This diff will remove the dependency and the ServiceName class. RegionName will be removed as a separate task since it is currently being used by other functions. ServiceName will be maintained in the billing_monitor.py in later diff.

Differential Revision: D31209369

fbshipit-source-id: 065a57e0fd0a67962e69331cbeed6a1244828f70