Add AWS services to ServiceName (#158)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/158

After offline discussion about the design, we decide that these two enum class ServiceName and RegionName is not a good way to keep up with the potential name change. This diff will remove the dependency and the ServiceName class. RegionName will be removed as a separate task since it is currently being used by other functions. ServiceName will be maintained in the billing_monitor.py in later diff.

Reviewed By: peking2

Differential Revision: D31209369

fbshipit-source-id: 292f9b890fb5b81498cf504a950e336d8aa5816c