Onedocker ignoring binary version (#165)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/165

The solutions pod is fighting for a Sev 3 S246342. The E2E test script told us that the PL rc binaries worked fine, so we promoted them to prod this afternoon, but turned out that the shard aggregator binary crashed. As we investigated why the E2E test didn't catch that, we noticed that this version parameter is ignored.

This diff won't resolve the sev but it's necessary to prevent it from happening again.

Reviewed By: peking2

Differential Revision: D31424129

fbshipit-source-id: 59b579c2a90203335685adb9cc8257aa0cd201cd