Update dependencies and release a new version (#207)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/207

fbpcp exposes a vulnerable due to the low version of some dependencies.
So I update the version of dependencies to make them internally consistent and Fix the docker image vulnerables

The root cause of the vulnerable is urllib3, which is used by boto3 is too old.

More context: T105173350

> This diff will not ship until the red zone ends

Reviewed By: peking2

Differential Revision: D32714430

fbshipit-source-id: 103178ea5352cb181855c1b3b4061f03300a29f0