Refactor the instance update logic in MPC service (#259)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/259

When mpc service queries the status of running containers, it read the container instance from its repository and using the container ids as the input to do the query.

It is very likely that part of the containers are cleaned up by AWS (1 hr ETA after stop), while others are still running, which makes the number of containerInstances inconsistent. We will need to refactor the logic of update MPC repository. Please see ``_get_updated_container`` for new logic

The new query logic:

1. If ECS returns an instance, return the instance to callers
2. If ECS could not find an instance and the status of existing instance is not stopped, return None to callers
3. If ECS cound not find an instance and the status of existing instance is stopped, return the existing instance to callers.

Reviewed By: taoyong-ty

Differential Revision: D34410047

fbshipit-source-id: 03dec4aa10686dcf5c34b2a56371332fb515ed4a