Handle different publisher VPC CIDR (#264)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/264

PCE Publisher and Partner mostly shares the same setup, as such the validation functions can be ran on both sides.  However, one difference need to address is VPC CIDR.  Partner VPC CIDR is always 10.1.0.0/16, while publisher is 10.0.0.0/16. This is hardcoded and during validation, I want to differiecieate publisher and partner by introducing a new argument: role.

To maintain the same behavior on partner end, if role is not specified, it will use partner by default.

Reviewed By: zehuali

Differential Revision: D34776172

fbshipit-source-id: b1fde584ae7c258a0271ee073fab3b8e915f910e