Fail validation when duplicate resources tagged with same pce:pce-id (#269)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/269

## Background
PCE Validator uniquely identifies a PCE (and the resources that make a PCE, read PCE Standard) by the PCE ID, a unique identifier which is stored on the resource in a cloud provider like AWS under a free-form tag value 'pce:pce-id'.

## Problem
Because tags and values are modifiable, and cloud providers like AWS do not enforce any uniqueness constraints on the tag values, it is easy to have situations with more than one resource is tagged with the same `pce:pce-id` value.
PCE Validator execution will lead to ambiguous results because of the duplicate resources now tagged with the same pce id.

## Solution
Instead of silently failing (by continuing to run validation suite against a PCE with multiple of resources that PCE Standard expects to be unique -- for eg VPC, route table), PCE Validator will now fail-fast when it encounters duplicate resources.

## In V6
The check for duplicate resources is done by a new class `pce.validator.duplicate_pce_resources_checker.DuplicatePCEResourcesChecker` *before* attempting to load a PCE via a PCEService::get_pce() interface, to keep the get_pce() interface clean and keep the concern of checking for duplicacy separately into the new class.
`DuplicatePCEResourcesChecker` is implemented in the preceding diff.

Reviewed By: yutong-w

Differential Revision: D34447727

fbshipit-source-id: 051f9be8ff30cb8955e2b3f8cf7d311e3e70f540