Refactor validator methods names to be derived from ValidatorStepNames (#286)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/286

To prepare for implementing --skip-step functionality to optionally skip some PCE validator steps, this diff introduces the "code_name" value in the ValidationStepNames enum, which is a lowercased, underscored name of the existing step name.

The `code_name` value is easier to enter on the command line as an argument to --skip-step, and is also the way most `validate_*` methods are named in the ValidationSuite class, corresponding to each ValidationStepNames.
This diff also renames the few validator methods that did not follow the pattern of being named "validate_{code_name}" for eg `validate_roles` -> `validate_iam_roles`

Reviewed By: Kewen12

Differential Revision: D35424006

fbshipit-source-id: 646b3dbb336ccbb4f7e3c6b38206c4017eb4334f