Add function to get list of files in Storage Service
Summary:
**Context:**
Currently we have a post processing handler in PID where it needs to read the all the metric files in a directory and compute intersection rate.

In order to get the names of all the metric files, we build all the possible paths that end with "_metrics" and check if each of the path is in the file system (s3). This is causing too many calls to s3 api. For example a PC run with 30 shards will make 30 calls to check if each of the metric paths exist. We have also seen a customer issue related to this.(related [post](https://fb.workplace.com/groups/331044242148818/permalink/516307276955846/))

**This diff:**
We add a new function to list files in the storage service that will help reduces the number of calls where we get the list of files in a directory.

Reviewed By: zhuang-93

Differential Revision: D37778726

fbshipit-source-id: d56e7914579069a46aa177e53f18f9d0e4d73adb