Remove Attestation service integration in onedocker runner (#395)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/395

As per team discussion, this attestation service is not ready to be OSS. Considering in the long term, Attestation will not be called separately when upload/download binary in onedocker_runner/onedocker_cli but through OneDocker Repository service. So we will removed all the integration in onedocker_runner and onedocker_cli (later diff).

Differential Revision: D38441652

fbshipit-source-id: 9a711c043e04fba372ab3380dec6a825347884c0