Remove Attestation service integration in onedocker runner (#395)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/395

As per team discussion, this attestation service is not ready to be OSS. Considering in the long term, Attestation will not be called separately when upload/download binary in onedocker_runner/onedocker_cli but through OneDocker Repository service. So we will removed all the integration in onedocker_runner and onedocker_cli (later diff).

Steps:
- remove Attestation service integration in onedocker runner (this diff)
- remove Attestation service integration in onedocker cli
- move Attestation service to internal folder
...

Reviewed By: YigeZhu

Differential Revision: D38441652

fbshipit-source-id: ba383efcfa6cc14db69ebae444aee3e952f293fd