Remove Attestation service integration in onedocker runner (#396)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/396

As per team discussion, this attestation service is not ready to be OSS. Considering in the long term, Attestation will not be called separately when upload/download binary in onedocker_runner/onedocker_cli but through OneDocker Repository service. So we will removed all the integration in onedocker_runner and onedocker_cli (later diff).
Steps:
- [done] remove Attestation service integration in onedocker runner
- [this diff] remove Attestation service integration in onedocker cli
move Attestation service to internal folder
...

Reviewed By: liliarizona

Differential Revision: D38467994

fbshipit-source-id: 9be7e3811a6277650bef36b19dadac611160ca98