short term solution to resolve backward compatibility
Summary:
**What's the problem?**
Onedocker runner and other services currently rely on a [hardcoded default version](https://fburl.com/code/tu8be5nx) to retrieve the binaries. With the new build and release process, binaries will no longer be versioned by stage names (i.e. "rc", "canary", "latest"), but will be given version numbers instead. Therefore, this might introduce breaks for advertiser runs that rely on the default value. This is currently a blocker for account migration in onedocker_runner.

**What does this diff do?**
This diff provides a way for the upload API to skip version validation check when the version is "latest". With this change, B&R will be able to upload to "latest" to ensure backward compatibility. We expect B&R to first call upload with a version number, then call upload with "latest" version at a proper stage.

**Why is this only a short term change and what's the long term solution?**
As I discussed with musebc, the long term solution would be for the build and release to provide a config with the prod version number that can be read during each run. The config will be updated during the build and release process to reflect the latest prod version number.

Differential Revision: D38745356

fbshipit-source-id: ccad2675a050307cb1f632c93d6829f6fbcf3518