Add validation support for partial config (#432)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/432

To address concerns in D38924284 (https://github.com/facebookresearch/fbpcp/commit/d7cffb395156cb4a530050a8e7b3721d3ec943eb)

Previously we build the services as long as it is present in the config. In this diff, we first examines whether the config is valid based on the command passed in and build the required services accordingly.

In the next diff, we will remove the services as global variable, and pass them locally to each command function.

Differential Revision: D39315955

fbshipit-source-id: 03561a51192d4a20a20d9b6ed5dbe1509b092aeb