UnOSS OneDocker core dump uploader and move it to Development (#437)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/437

This OneDocker core dump uploader is pending for PXFN and will be removed from prod until it is done. For now to unblock OneDocker image release, we will move this to internal folder. Due to the amount of files/code change, this is broken down to 2 diffs.

This diff is to remove the OSS code.

Differential Revision: D39866707

fbshipit-source-id: 2e87e46c2a71310ea0b0a4bcebd0a11d065dc5ac