UnOSS OneDocker Self-signed TLS and move it to Development
Summary: Based on the guidances from Security partner, this self-signed feature should only be used for deployment. Currently we only have one OneDocker runner that shared between prod and dev. To unblock OneDocker image release, for now we will unoss this feature and move it to internal. In the long term, we plan to refactor onedocker-runner to support plug in features so it can be easily enable/disable based on run enviroment.

Differential Revision: D39896486

fbshipit-source-id: 95d055e67fa8abc4aef17127d50e5f88b2b27e74