UnOSS OneDocker Self-signed TLS and move it to Development (#438)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/438

Based on the guidances from Security partner, this self-signed feature should only be used for deployment. Currently we only have one OneDocker runner that shared between prod and dev. To unblock OneDocker image release, for now we will unoss this feature and move it to internal. In the long term, we plan to refactor onedocker-runner to support plug in features so it can be easily enable/disable based on run enviroment.

Reviewed By: YigeZhu

Differential Revision: D39896486

fbshipit-source-id: d2bb0f65ac72f85949fd870898a1eb656262bb0b