Improve the mapping for peering status (#450)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/450

The current mapping is not sufficient to describe the immediate vpc peering status from a related API call. And the `provisioning` is mapped to `NOT_READY` with `failed`, `deleted`, `deleting`, and `expire` which complicates the logic to check the success of the peering. The `provisioning` status after acceptance only lasts for a very short period (test: D40107684) but it is a state prior to `active` indicating the acceptance is successful. Therefore, we separated the `provisioning` and use `inactive` to represent the 'unsuccessful' status.

The new mapping:

AWS:
```
Pending-acceptance -> pending_acceptance
Provisioning -> not_ready
Active -> active
Failed, expired, rejected, deleted, deleting -> inactive
Initiating-request -> initiating
```

GCP
```
Active -> active
Else -> inactive
```

{F778655190}

Ref: https://fburl.com/2oubduyj (AWS)
https://fburl.com/bmpd74ex (GCP)

Reviewed By: ajaybhargavb

Differential Revision: D40152913

fbshipit-source-id: 956a80ad7b2f2d9433395873e881efb6bf0d201e