(Simple) UnOSS KeyManagementService
Summary: Refer to D38441652 (https://github.com/facebookresearch/fbpcp/commit/8090186509cce01abfa6a3b514495c944f7bf0b9) for more context. This KeyManagementService is a stretch goal for the previous intern project. Currently it is not used anywhere. UnOSS and move it to intern folder for now for best practice.

Differential Revision: D40536607

fbshipit-source-id: 142fc0ea481f958f6de96e4f4c313e6b1fe755be