(Simple) UnOSS KeyManagementService (#453)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/453

Refer to D38441652 (https://github.com/facebookresearch/fbpcp/commit/8090186509cce01abfa6a3b514495c944f7bf0b9) for more context. This KeyManagementService is a stretch goal for the previous intern project. Currently it is not used anywhere. UnOSS and move it to intern folder for now for best practice.

Reviewed By: Kewen12

Differential Revision: D40536607

fbshipit-source-id: 5af7cc52a2ba03a412e6c35c5e5536c748be8660