add server_uris to MPCInstance
Summary:
To establish TLS connection, we need to route each container's ip_address to a server_hostname/server_uri. Each container should be assigned with a unique server uris.

Additional context:
Certificates are generated with the domain of this server_uris and supports multiple subdomains. We need a different address for every node because eventually we will expect the nodes to be looked up via DNS.

Differential Revision: D40917008

fbshipit-source-id: 5dcd558d923235af5635023bfbd171c9927e9aad