OSS OneDocker metadata service
Summary:
As implementation and integration tests are done, this diff is to move metadata service to OSS folder. This won't be release until H1 2023. But we need to move this to OSS in order to integrate this feature to OneDocker repository service.

I'm in the process of requesting exception for this change. Will land only when diff and exception are both approved

Differential Revision: D41481063

fbshipit-source-id: 6bde7cf254123adc12a4bf2440bd9adb96cde277