OSS OneDocker metadata service (#466)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/466

As implementation and integration tests are done, this diff is to move metadata service to OSS folder. This won't be release until H1 2023. But we need to move this to OSS in order to integrate this feature to OneDocker repository service.

I'm in the process of requesting exception for this change. Will land only when diff and exception are both approved

Reviewed By: peking2, Kewen12

Differential Revision: D41481063

fbshipit-source-id: fa6acf462af61821c34dc790f12250ffdf496665