Remove invalid S3 bucket from unit tests - OSS (#485)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/485

The S3 buckets used from onedocker runner unit test are out dated and has been taken over externally. Although these S3 buckets are not actually accessed in the unitest as the connection are mocked. These usage still raise whitehat security task. To resolve this issue and avoid triggering similar tasks in the future, this diff is to replace the invalid buckets with mock S3 path.

This diff is to change the OSS unittest with changelog change.

1. Type of changes
-modify

2. Requires an OSS plan：
-no

3. OSS plan link or explain why the change doesn't require an OSS plan:
-minor change to unittest

4. Have you added change description to [CHANGELOG.md](https://www.internalfb.com/code/fbsource/fbcode/measurement/private_measurement/pcp/oss/CHANGELOG.md)? Please provide diff # if not in this diff.
-yes

5. Is it covered by unit test, please provide a link/diff if test is not included in this diff
-yes

6. Is it covered by integration test, please provide a link/diff if test is not included in this diff
-no

Reviewed By: ajaybhargavb

Differential Revision: D43511094

fbshipit-source-id: 3fc900b4522021ea7aacce0a1593e670066c453e