UnOSS OneDocker cryptogateway and remove fbpcp dependency to cryptography package (#490)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/490

OSS auto checkup tool has detected a security vulnerability for the version of cryptography that we are using in OSS. This version was updated in D37078211 (https://github.com/facebookresearch/fbpcp/commit/2b3b4f15fd94268d6e6c1010e12aed891dcf0ad9) to match internal version to avoid inconsistency. As this cryptography package is only used for CryptographyGateway, which will be decoupled with OneDocker once we migrate to OPA. For now we will just unoss this CryptographyGateway to remove dependency to cryptography package.

1. Type of changes
remove
2. Requires an OSS plan
No
3. OSS plan link or explain why the change doesn't require an OSS plan
Remove unused feature
4. Have you added change description to [CHANGELOG.md](https://www.internalfb.com/code/fbsource/fbcode/measurement/private_measurement/pcp/oss/CHANGELOG.md)?
Please provide diff # if not in this diff.
Yes
5. Is it covered by unit test, please provide a link/diff if test is not included in this diff
yes
6. Is it covered by integration test, please provide a link/diff if test is not included in this diff
No (will add integration test later when this feature is utilized again)

Reviewed By: yutong-w

Differential Revision: D43546751

fbshipit-source-id: 0dce85ce8e85a60e9c5e73a7cc7d777cfc2deaf2