Update ContainerService to accept a permissions arg (#509)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcp/pull/509

This change updates the Container Service to take a permissions argument when creating instances. This argument will configure the permission for the container created. The Container Instance returned from Container Service will also now include a property describing the container's permissions, when overridden from the default.

The permissions will be used in a later diff to configure permissions allowing the container to access TLS secret/private key, when the feature is enabled.

Currently, the permission argument will only drive behavior for AWS-ECS Container Service.

Reviewed By: jliu0501

Differential Revision: D44140791

fbshipit-source-id: e8772c5ec96e5ae86fd8d3d13349cbeb25d012f2